requirements.txt
pandas
openpyxl
boto3
python-dotenv
requests
apache-airflow-providers-amazon
apache-airflow-providers-smtp>=1.7.0
